fix(cloud-native): inconsistent env naming used for modifying jans-config-api configuration
