+++ docker-jans-config-api/scripts/bootstrap.py
@@ -12,7 +12,6 @@ from ldif import LDIFWriter
 from jans.pycloudlib import get_manager
 from jans.pycloudlib import wait_for_persistence
 from jans.pycloudlib.persistence.hybrid import render_hybrid_properties
-from jans.pycloudlib.persistence.sql import doc_id_from_dn
 from jans.pycloudlib.persistence.sql import SqlClient
 from jans.pycloudlib.persistence.sql import render_sql_properties
 from jans.pycloudlib.persistence.sql import override_simple_json_property
@@ -35,6 +34,7 @@ from utils import get_config_api_scope_mapping
 from utils import get_ads_project_base64
 from utils import AUI_AGAMA_PW_ARCHIVE
 from utils import AUI_AGAMA_PW_DEPLOYMENT_ID
+from utils import URLModifier
 
 logging.config.dictConfig(LOGGING_CONFIG)
 logger = logging.getLogger("jans-config-api")
@@ -282,40 +282,7 @@ class PersistenceSetup:
         self.client = client_cls(manager)
 
         self.plugins = plugins or []
-
-    def get_auth_config(self):
-        dn = "ou=jans-auth,ou=configuration,o=jans"
-        entry = self.client.get("jansAppConf", doc_id_from_dn(dn))
-        return json.loads(entry["jansConfDyn"])
-
-    def transform_url(self, url):
-        auth_server_url = os.environ.get("CN_AUTH_SERVER_URL", "")
-
-        if not auth_server_url:
-            return url
-
-        parse_result = urlparse(url)
-        if parse_result.path.startswith("/.well-known"):
-            path = f"/jans-auth{parse_result.path}"
-        else:
-            path = parse_result.path
-        return f"http://{auth_server_url}{path}"
-
-    def get_injected_urls(self):
-        auth_config = self.get_auth_config()
-
-        urls = (
-            "issuer",
-            "openIdConfigurationEndpoint",
-            "introspectionEndpoint",
-            "tokenEndpoint",
-            "tokenRevocationEndpoint",
-        )
-
-        return {
-            url: self.transform_url(auth_config[url])
-            for url in urls
-        }
+        self.url_modifier = URLModifier(self.client)
 
     @cached_property
     def ctx(self) -> dict[str, _t.Any]:
@@ -335,7 +302,7 @@ class PersistenceSetup:
             "endpointInjectionEnabled": "true",
             "configOauthEnabled": str(os.environ.get("CN_CONFIG_API_OAUTH_ENABLED") or True).lower(),
         }
-        ctx.update(self.get_injected_urls())
+        ctx.update(self.url_modifier.get_injected_urls())
 
         # Client
         ctx["jca_client_id"] = self.manager.config.get("jca_client_id")
